Started to move buisenes logic to the container
