+++ app/components/PredictionMarkets/PredictionMarkets.jsx
@@ -14,35 +14,21 @@ import {ChainStore, FetchChainObjects} from "bitsharesjs";
 import {Switch, Button, Radio, Icon, Tooltip} from "bitshares-ui-style-guide";
 import {Asset, Price} from "../../lib/common/MarketClasses";
 import Translate from "react-translate-component";
-import {bindToCurrentAccount} from "../Utility/BindToCurrentAccount";
-import AssetStore from "../../stores/AssetStore";
-import MarketsStore from "../../stores/MarketsStore";
-import {connect} from "alt-react";
-import {getPredictionMarketIssuers} from "../../lib/chain/onChainConfig";
 
 class PredictionMarkets extends Component {
     constructor(props) {
         super(props);
         this.state = {
-            lastAssetSymbol: null,
-            predictionMarkets: [],
-            isFetchingFinished: false,
             searchTerm: "",
             detailsSearchTerm: "",
-            selectedPredictionMarket: null,
-            opinions: [],
-            fetchedAssets: [],
             preselectedOpinion: "yes",
             preselectedAmount: 0,
             preselectedProbability: 0,
             isCreateMarketModalOpen: false,
             isAddOpinionModalOpen: false,
             isResolveModalOpen: false,
-            isHideUnknownHousesChecked: true,
-            isHideInvalidAssetsChecked: true,
             opinionFilter: "yes",
-            predictionMarketAssetFilter: "open",
-            whitelistedHouses: []
+            predictionMarketAssetFilter: "open"
         };
 
         this.onCreatePredictionMarketModalOpen = this.onCreatePredictionMarketModalOpen.bind(
@@ -59,314 +45,12 @@ class PredictionMarkets extends Component {
         this.onResolveModalOpen = this.onResolveModalOpen.bind(this);
         this.onResolveModalClose = this.onResolveModalClose.bind(this);
         this.updateAsset = this.updateAsset.bind(this);
-        this.handleUnknownHousesToggleChange = this.handleUnknownHousesToggleChange.bind(
-            this
-        );
-        this.handleInvalidAssetsChecked = this.handleInvalidAssetsChecked.bind(
-            this
-        );
-    }
-
-    componentWillMount() {
-        this.getWhitelistedHousesThroughAsset();
-        //    this._checkAssets(this.props.assets);
-    }
-
-    componentWillReceiveProps(np) {
-        if (np.assets !== this.props.assets) {
-            this._checkAssets(np.assets);
-        }
-
-        if (np.marketLimitOrders !== this.props.marketLimitOrders) {
-            this._updateOpinionsList(np.marketLimitOrders);
-        }
-    }
-
-    async getWhitelistedHousesThroughAsset() {
-        let whitelistedHouses = await getPredictionMarketIssuers();
-        whitelistedHouses = ["1.2.428447", "1.2.1099493", "1.2.160399"]; //!!!!!!!!!FOR TESTING!!!!!!!!!!!!!!!!!
-        this.setState({
-            whitelistedHouses
-        });
-        this._getWhitelistedAssets(whitelistedHouses);
-    }
-
-    async _getWhitelistedAssets(whitelistedHouses) {
-        let assets = [];
-        let accountObjects = await FetchChainObjects(
-            ChainStore.getAsset,
-            whitelistedHouses,
-            undefined,
-            {}
-        );
-        accountObjects.forEach(item => {
-            if (item) {
-                item = item.toJS();
-                assets = [...assets, ...item.assets];
-            }
-        });
-        let assetsObjects = await FetchChainObjects(
-            ChainStore.getAsset,
-            assets,
-            undefined,
-            {}
-        );
-        assetsObjects = assetsObjects.map(item => item.toJS());
-
-        console.log("assetsObjects", assetsObjects);
-    }
-
-    _checkAssets(fetchedAssets) {
-        let searchAsset = this.state.lastAssetSymbol
-            ? this.state.lastAssetSymbol
-            : "A";
-
-        if (fetchedAssets) {
-            const lastAsset = fetchedAssets
-                .sort((a, b) => {
-                    if (a.symbol > b.symbol) {
-                        return 1;
-                    } else if (a.symbol < b.symbol) {
-                        return -1;
-                    } else {
-                        return 0;
-                    }
-                })
-                .last();
-            searchAsset = lastAsset ? lastAsset.symbol : "A";
-
-            // parse flags and description
-            fetchedAssets.forEach(item => {
-                if (!item.forPredictions) {
-                    item.forPredictions = {
-                        description: assetUtils.parseDescription(
-                            item.options.description
-                        ),
-                        flagBooleans: assetUtils.getFlagBooleans(
-                            item.options.flags,
-                            true
-                        )
-                    };
-                }
-            });
-
-            this._updatePredictionMarketsList(fetchedAssets);
-        }
-        if (
-            !this.state.lastAssetSymbol ||
-            this.state.lastAssetSymbol !== searchAsset
-        ) {
-            AssetActions.getAssetList.defer(searchAsset, 100);
-            this.setState({
-                lastAssetSymbol: searchAsset
-            });
-        } else {
-            this.setState({
-                isFetchingFinished: true,
-                fetchedAssets
-            });
-        }
-    }
-
-    _isKnownIssuer(asset) {
-        return this.state.whitelistedHouses.includes(asset.issuer);
-    }
-
-    _isValidPredictionMarketAsset(asset) {
-        // must have valid date
-        const resolutionDate = new Date(
-            asset.forPredictions.description.expiry
-        );
-        if (resolutionDate instanceof Date && isNaN(resolutionDate.getTime())) {
-            return false;
-        }
-        // must have description and prediction filled
-        if (!asset.forPredictions.description.condition) {
-            return false;
-        }
-        if (!asset.forPredictions.description.main) {
-            return false;
-        }
-        // must have meaningfull description and prediction
-        if (asset.forPredictions.description.condition.length < 10) {
-            return false;
-        }
-        if (asset.forPredictions.description.main.length < 20) {
-            return false;
-        }
-        // market fee may not be crazy
-        if (asset.options.market_fee_percent / 100 >= 10) {
-            return false;
-        }
-        return true;
-    }
-
-    _updatePredictionMarketsList(fetchedAssets = null) {
-        if (fetchedAssets == null) {
-            fetchedAssets = this.state.fetchedAssets;
-        }
-        const filter = this.state.predictionMarketAssetFilter;
-        const assets = fetchedAssets.filter(asset => {
-            if (
-                asset.bitasset_data &&
-                asset.bitasset_data.is_prediction_market
-            ) {
-                if (
-                    this.state.isHideUnknownHousesChecked &&
-                    !this._isKnownIssuer(asset)
-                ) {
-                    return false;
-                } else if (
-                    this.state.isHideInvalidAssetsChecked &&
-                    !this._isValidPredictionMarketAsset(asset)
-                ) {
-                    return false;
-                } else {
-                    if (filter && !(filter === "all")) {
-                        const resolutionDate = new Date(
-                            asset.forPredictions.description.expiry
-                        );
-                        const isExpiredOrResolved =
-                            asset.bitasset_data.settlement_fund > 0 ||
-                            resolutionDate < new Date();
-                        if (filter === "open") {
-                            return !isExpiredOrResolved;
-                        } else if (filter === "past_resolution_date") {
-                            return isExpiredOrResolved;
-                        } else {
-                            return false;
-                        }
-                    } else {
-                        return true;
-                    }
-                }
-            } else {
-                return false;
-            }
-        });
-        let predictionMarkets = [...assets].map(item => {
-            let market_fee = 0;
-            let max_market_fee = 0;
-            if (item[1].forPredictions.flagBooleans["charge_market_fee"]) {
-                market_fee = item[1].options.market_fee_percent;
-                max_market_fee = item[1].options.max_market_fee;
-            }
-            let predictionMarketTableRow = {
-                asset: item,
-                asset_id: item[1].id,
-                issuer: item[1].issuer,
-                description: item[1].forPredictions.description.main,
-                symbol: item[1].symbol,
-                condition: item[1].forPredictions.description.condition,
-                expiry: item[1].forPredictions.description.expiry,
-                options: item[1].options,
-                marketConfidence: 0,
-                marketLikelihood: 0,
-                market_fee,
-                max_market_fee
-            };
-            return predictionMarketTableRow;
-        });
-
-        this.setState({
-            predictionMarkets
-        });
-    }
-
-    _updateOpinionsList(fetchedOpinions) {
-        let orders = [];
-        const selectedMarket = this.state.selectedPredictionMarket;
-        fetchedOpinions.forEach((order, order_id) => {
-            const opinion =
-                order.market_base === order.sell_price.base.asset_id
-                    ? "no"
-                    : "yes";
-            const refPrice =
-                order.market_base === order.sell_price.base.asset_id
-                    ? order.sell_price.invert().toReal()
-                    : order.sell_price.toReal();
-            const amount =
-                order.market_base === order.sell_price.base.asset_id
-                    ? order.amountForSale()
-                    : order.amountToReceive();
-            const premium =
-                order.market_base === order.sell_price.base.asset_id
-                    ? order.amountToReceive()
-                    : order.amountForSale();
-            const flagBooleans = assetUtils.getFlagBooleans(
-                selectedMarket.options.flags,
-                true
-            );
-            let fee = 0;
-            if (flagBooleans["charge_market_fee"]) {
-                fee = Math.min(
-                    selectedMarket.options.max_market_fee,
-                    (amount.amount *
-                        selectedMarket.options.market_fee_percent) /
-                        10000
-                );
-            }
-
-            if (refPrice < 1) {
-                orders.push({
-                    order_id,
-                    opinionator: order.seller,
-                    opinion,
-                    amount,
-                    likelihood: refPrice,
-                    potentialProfit: new Asset({
-                        amount: amount.amount,
-                        asset_id: premium.asset_id,
-                        precision: premium.precision
-                    }),
-                    premium: premium,
-                    commission: new Asset({
-                        amount: fee * refPrice,
-                        asset_id: premium.asset_id,
-                        precision: premium.precision
-                    })
-                });
-            }
-        });
-        this.setState({opinions: [...orders]});
-    }
-
-    async getMarketOpinions(market) {
-        if (this.state.subscribedMarket) {
-            await MarketsActions.unSubscribeMarket(
-                this.state.subscribedMarket.quote.get("id"),
-                this.state.subscribedMarket.base.get("id")
-            );
-        }
-        const base = ChainStore.getAsset(
-            market.options.core_exchange_rate.base.asset_id
-        );
-        const quote = ChainStore.getAsset(
-            market.options.core_exchange_rate.quote.asset_id
-        );
-        await MarketsActions.subscribeMarket(
-            base,
-            quote,
-            this.props.bucketSize,
-            this.props.currentGroupOrderLimit
-        );
-        this.setState({
-            subscribedMarket: {
-                base,
-                quote
-            }
-        });
     }
 
     onMarketAction({market, action}) {
+        this.props.onMarketAction({market, action});
         if (typeof action === "string") {
             //on buttons action
-            if (!this.state.selectedPredictionMarket) {
-                this.setState({
-                    selectedPredictionMarket: market
-                });
-            }
-
             switch (action) {
                 case "resolve": {
                     this.setState({
@@ -404,20 +88,6 @@ class PredictionMarkets extends Component {
                         preselectedProbability: 0
                     });
                 }
-            }
-        } else {
-            //on row action
-            if (this.state.selectedPredictionMarket) {
-                this.setState({
-                    selectedPredictionMarket: null
-                });
-            } else {
-                this.setState(
-                    {
-                        selectedPredictionMarket: market
-                    },
-                    () => this.getMarketOpinions(market)
-                );
             }
         }
     }
@@ -471,30 +141,6 @@ class PredictionMarkets extends Component {
         this.setState({
             isResolveModalOpen: false
         });
-    }
-
-    handleUnknownHousesToggleChange() {
-        const isHideUnknownHousesChecked = !this.state
-            .isHideUnknownHousesChecked;
-        this.setState(
-            {
-                isHideUnknownHousesChecked,
-                selectedPredictionMarket: null
-            },
-            () => this._updatePredictionMarketsList()
-        );
-    }
-
-    handleInvalidAssetsChecked() {
-        const isHideInvalidAssetsChecked = !this.state
-            .isHideInvalidAssetsChecked;
-        this.setState(
-            {
-                isHideInvalidAssetsChecked,
-                selectedPredictionMarket: null
-            },
-            () => this._updatePredictionMarketsList()
-        );
     }
 
     onOppose = opinion => {
@@ -571,14 +217,14 @@ class PredictionMarkets extends Component {
     }
 
     getOverviewSection() {
-        const setPredictionMarketAssetFilter = e => {
-            this.setState(
-                {
-                    predictionMarketAssetFilter: e.target.value
-                },
-                this._updatePredictionMarketsList
-            );
-        };
+        // const setPredictionMarketAssetFilter = e => {
+        //     this.setState(
+        //         {
+        //             predictionMarketAssetFilter: e.target.value
+        //         },
+        //         this._updatePredictionMarketsList
+        //     );
+        // };
         return (
             <div>
                 <div
@@ -593,7 +239,7 @@ class PredictionMarkets extends Component {
                         <Radio.Group
                             style={{marginLeft: "20px"}}
                             value={this.state.predictionMarketAssetFilter}
-                            onChange={setPredictionMarketAssetFilter}
+                            //                     onChange={setPredictionMarketAssetFilter}
                         >
                             <Radio value={"all"}>
                                 {counterpart.translate(
@@ -691,16 +337,16 @@ class PredictionMarkets extends Component {
                     </span>
                 </div>
                 <PredictionMarketsOverviewTable
-                    predictionMarkets={this.state.predictionMarkets}
+                    predictionMarkets={this.props.predictionMarkets}
                     currentAccount={this.props.currentAccount}
                     onMarketAction={this.onMarketAction}
                     searchTerm={this.state.searchTerm.toUpperCase()}
                     selectedPredictionMarket={
-                        this.state.selectedPredictionMarket
+                        this.props.selectedPredictionMarket
                     }
                     hideUnknownHouses={this.state.isHideUnknownHousesChecked}
-                    loading={!this.state.isFetchingFinished}
-                    whitelistedHouses={this.state.whitelistedHouses}
+                    loading={!this.props.isFetchingFinished}
+                    whitelistedHouses={this.props.whitelistedHouses}
                 />
             </div>
         );
@@ -786,12 +432,12 @@ class PredictionMarkets extends Component {
                         </Button>
                     </span>
                 </div>
-                {this.state.opinions ? (
+                {this.props.opinions ? (
                     <PredictionMarketDetailsTable
                         predictionMarketData={{
-                            predictionMarket: this.state
+                            predictionMarket: this.props
                                 .selectedPredictionMarket,
-                            opinions: this.state.opinions
+                            opinions: this.props.opinions
                         }}
                         currentAccount={this.props.currentAccount}
                         onOppose={this.onOppose}
@@ -806,7 +452,7 @@ class PredictionMarkets extends Component {
 
     render() {
         const symbols = [...this.props.assets].map(item => item[1].symbol);
-
+        console.log(this.state);
         return (
             <div
                 className="prediction-markets grid-block vertical"
@@ -819,7 +465,7 @@ class PredictionMarkets extends Component {
                     <HelpContent path={"components/PredictionMarkets"} />
                 </div>
                 {this.getOverviewSection()}
-                {this.state.selectedPredictionMarket
+                {this.props.selectedPredictionMarket
                     ? this.getDetailsSection()
                     : null}
                 {this.state.isCreateMarketModalOpen ? (
@@ -835,7 +481,7 @@ class PredictionMarkets extends Component {
                     <AddOpinionModal
                         visible={this.state.isAddOpinionModalOpen}
                         onClose={this.onAddOpinionModalClose}
-                        predictionMarket={this.state.selectedPredictionMarket}
+                        predictionMarket={this.props.selectedPredictionMarket}
                         opinion={this.state.initialOpinion}
                         currentAccount={this.props.currentAccount}
                         submitNewOpinion={this.onSubmitNewOpinion}
@@ -844,15 +490,15 @@ class PredictionMarkets extends Component {
                         preselectedProbability={
                             this.state.preselectedProbability
                         }
-                        baseAsset={this.state.subscribedMarket.base}
-                        quoteAsset={this.state.subscribedMarket.quote}
+                        baseAsset={this.props.subscribedMarket.base}
+                        quoteAsset={this.props.subscribedMarket.quote}
                     />
                 ) : null}
                 {this.state.isResolveModalOpen ? (
                     <ResolveModal
                         visible={this.state.isResolveModalOpen}
                         onClose={this.onResolveModalClose}
-                        predictionMarket={this.state.selectedPredictionMarket}
+                        predictionMarket={this.props.selectedPredictionMarket}
                         onResolveMarket={this.onResolveMarket}
                     />
                 ) : null}
@@ -861,23 +507,4 @@ class PredictionMarkets extends Component {
     }
 }
 
-PredictionMarkets = connect(
-    PredictionMarkets,
-    {
-        listenTo() {
-            return [AssetStore, MarketsStore];
-        },
-        getProps() {
-            return {
-                assets: AssetStore.getState().assets,
-                markets: MarketsStore.getState().marketData,
-                bucketSize: MarketsStore.getState().bucketSize,
-                currentGroupOrderLimit: MarketsStore.getState()
-                    .currentGroupLimit,
-                marketLimitOrders: MarketsStore.getState().marketLimitOrders
-            };
-        }
-    }
-);
-
-export default (PredictionMarkets = bindToCurrentAccount(PredictionMarkets));
+export default PredictionMarkets;
